pyproject.toml
[tool.poetry]
name = "tzelleke/eda-service-aws"
version = "0.1.0"
description = ""
authors = ["tzelleke <it@tzelleke.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.70.0"
constructs = "^10.0.0"
cdk-dynamo-table-view = "^0.2.486"
chalice = {extras = ["cdkv2"], version = "^1.29.0"}


[tool.poetry.group.tests.dependencies]
pytest = "^7.2.2"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.278"


[tool.black]
extend-exclude = '''
/(
    \.idea
)/
'''
include = '\.pyi?$'
line-length = 88
target-version = ["py310"]


[tool.ruff]
cache-dir = "~/.cache/ruff"
extend-exclude = []
extend-select = ["I"]
line-length = 88
target-version = "py311"


[tool.ruff.isort]
force-sort-within-sections = true


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
